test_samplecode.py
import unittest
import samplecode

class TestSamplecode(unittest.TestCase):
    def test_add(self):
        # TODO: Implement test for add
        # Expected return value of type int
        result = samplecode.add(a=None, b=None)
        self.assertEqual(result, None)  # Replace None with the expected result

    def test_subtract(self):
        # TODO: Implement test for subtract
        # Expected return value of type float
        result = samplecode.subtract(a=None, b=None)
        self.assertEqual(result, None)  # Replace None with the expected result

    def test_multiply(self):
        # TODO: Implement test for multiply
        # Expected return value of type int
        result = samplecode.multiply(a=None, b=None, c=None)
        self.assertEqual(result, None)  # Replace None with the expected result

    def test_divide(self):
        # TODO: Implement test for divide
        # Expected return value of type float
        result = samplecode.divide(self=None, a=None, b=None)
        self.assertEqual(result, None)  # Replace None with the expected result

    def test_power(self):
        # TODO: Implement test for power
        # Expected return value of type int
        result = samplecode.power(self=None, base=None, exponent=None)
        self.assertEqual(result, None)  # Replace None with the expected result

    def test_square_root(self):
        # TODO: Implement test for square_root
        # Expected return value of type float
        result = samplecode.square_root(self=None, number=None)
        self.assertEqual(result, None)  # Replace None with the expected result

    def test_Calculator_multiply(self):
        # TODO: Implement test for Calculator.multiply
        # Expected return value of type int
        obj = samplecode.Calculator()  # Create an instance of the class
        result = obj.multiply(a=None, b=None)
        self.assertEqual(result, None)  # Replace None with the expected result

    def test_Calculator_divide(self):
        # TODO: Implement test for Calculator.divide
        # Expected return value of type float
        obj = samplecode.Calculator()  # Create an instance of the class
        result = obj.divide(a=None, b=None)
        self.assertEqual(result, None)  # Replace None with the expected result

    def test_ScientificCalculator_power(self):
        # TODO: Implement test for ScientificCalculator.power
        # Expected return value of type int
        obj = samplecode.ScientificCalculator()  # Create an instance of the class
        result = obj.power(base=None, exponent=None)
        self.assertEqual(result, None)  # Replace None with the expected result

    def test_ScientificCalculator_square_root(self):
        # TODO: Implement test for ScientificCalculator.square_root
        # Expected return value of type float
        obj = samplecode.ScientificCalculator()  # Create an instance of the class
        result = obj.square_root(number=None)
        self.assertEqual(result, None)  # Replace None with the expected result


if __name__ == "__main__":
    unittest.main()
